impr: Refactor directories
