+++ ryot_compass/src/bevy_compass/drawing/draw.rs
@@ -1,5 +1,4 @@
-use crate::bevy_compass::drawing::brush::{Brush, RoundBrush};
-use crate::{Cursor, DrawingAction};
+use crate::{Brush, Cursor, DrawingAction, SquareBrush};
 use bevy::ecs::system::EntityCommand;
 use bevy::prelude::*;
 use leafwing_input_manager::prelude::*;
@@ -51,7 +50,7 @@ pub(super) fn draw_to_tile<C: ContentAssets>(
         let layer = prepared_appearance.layer;
         let appearance = AppearanceDescriptor::new(*group, *id, default());
 
-        let bundles = RoundBrush::new(5).to_paint(DrawingBundle::new(layer, *tile_pos, appearance));
+        let bundles = SquareBrush::new(5).apply(DrawingBundle::new(layer, *tile_pos, appearance));
         let mut queued = 0;
 
         for new_bundle in bundles {
